move `get_config_entry` to services module
