+++ homeassistant/components/habitica/util.py
@@ -22,13 +22,8 @@ from dateutil.rrule import (
 
 from homeassistant.components.automation import automations_with_entity
 from homeassistant.components.script import scripts_with_entity
-from homeassistant.config_entries import ConfigEntryState
 from homeassistant.core import HomeAssistant
-from homeassistant.exceptions import ServiceValidationError
 from homeassistant.util import dt as dt_util
-
-from .const import DOMAIN
-from .types import HabiticaConfigEntry
 
 
 def next_due_date(task: dict[str, Any], last_cron: str) -> datetime.date | None:
@@ -144,18 +139,3 @@ def get_recurrence_rule(recurrence: rrule) -> str:
 
     """
     return str(recurrence).split("RRULE:")[1]
-
-
-def get_config_entry(hass: HomeAssistant, entry_id: str) -> HabiticaConfigEntry:
-    """Return config entry or raise if not found or not loaded."""
-    if not (entry := hass.config_entries.async_get_entry(entry_id)):
-        raise ServiceValidationError(
-            translation_domain=DOMAIN,
-            translation_key="entry_not_found",
-        )
-    if entry.state is not ConfigEntryState.LOADED:
-        raise ServiceValidationError(
-            translation_domain=DOMAIN,
-            translation_key="entry_not_loaded",
-        )
-    return entry
